Horreum file re-org
From → To:
lib/familia/horreum/
├── core/
│   ├── connection.rb           → connection.rb
│   ├── database_commands.rb    → database_commands.rb
│   ├── serialization.rb        → persistence.rb, serialization.rb
│   └── utils.rb                → utils.rb
├── subclass/
│   ├── definition.rb           → definition.rb
│   ├── management.rb           → management.rb
│   └── related_fields_management.rb → relationships.rb (renamed)
├── shared/
│   └── settings.rb             → settings.rb
└── core.rb                     → DELETE (no longer needed)
