+++ lib/familia/horreum/connection.rb
