+++ try/features/transient_fields/refresh_reset_try.rb
@@ -8,6 +8,8 @@ Familia.debug = false
 Familia.dbclient.flushdb
 
 class SecretService < Familia::Horreum
+  feature :transient_fields
+
   identifier_field :name
 
   field :name
